+++ src/pyqasm/accelerate/linalg.pyx
